backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
slowapi==0.1.9
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.8
httpx==0.26.0
Pillow==10.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
aiobotocore[boto3]==2.12.3
requests
redis==5.0.1
websockets==12.0
stripe==8.0.0
authlib==1.3.0
httpx==0.26.0
sib-api-v3-sdk==7.6.0
google-api-python-client==2.111.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
